requirements.txt
# torch
# torchvision
# torchaudio
numpy
numba
torchtyping
tqdm
scipy
sklearn
librosa
pyyaml
PyWavelets
pyworld
pyopenjtalk
onnxruntime
